feat: Added masterArgsWait config option
Allows configuration of how long the workers will wait for the master process to send the arguments back to the worker. Useful for when there is lots of IO blocking work or promise resolution chains which prevents the master from responding within the 5 seconds allowed.

This is just a release message for #2